Add ScopedServiceFactory for consumer messages. Add repository project for example.

+++ RabbitDemo.Producer/Model/Dto/OrderMessage.cs
@@ -4,7 +4,7 @@ namespace RabbitDemo.Producer.Model.Dto;
 
 public class OrderMessage : IMessage
 {
-    public long OrderId { get; set; }
+    public int OrderId { get; set; }
     public string Name { get; set; }
     public string Description { get; set; }
     public int ClientId { get; set; }
